src/app/backend/pyproject.toml
# src/app/backend/pyproject.toml
[project]
name = "ccny-agent-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.1",
  "uvicorn[standard]>=0.35.0",
  "langgraph>=0.2.36",
  "langchain-openai>=0.1.21",
  "pydantic>=2.8.2",
  "python-dotenv>=1.0.1",
  "firecrawl-py>=0.0.15",
  "requests>=2.32.3",
  "openai>=1.40.0"
]

[tool.uv]
managed = true
